+++ app/src/main/java/com/example/parcialtp3/fragments/AutosFragment.kt
@@ -5,18 +5,21 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.example.parcialtp3.R
 import com.example.parcialtp3.adapters.CarsAdapter
-import com.example.parcialtp3.handlers.CarServiceHandler
+import com.example.parcialtp3.apiServiceBuilder.APIServicesBuilder
+import com.example.parcialtp3.model.domain.Car
 import com.example.parcialtp3.ui.viewmodel.AutosViewModel
+import retrofit2.Call
+import retrofit2.Callback
+import retrofit2.Response
 
 class AutosFragment : Fragment() {
     lateinit var autosView : View
     lateinit var autosRecView : RecyclerView
-    lateinit var autosViewModel : AutosViewModel
+    //lateinit var autosViewModel : AutosViewModel
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -30,19 +33,37 @@ class AutosFragment : Fragment() {
 
         autosRecView = autosView.findViewById<RecyclerView>(R.id.carsRecycleView)
         autosRecView.setHasFixedSize(true)
-        autosViewModel = ViewModelProvider(requireActivity()).get(AutosViewModel::class.java)
 
-        autosViewModel.createCarList(CarServiceHandler())
-
-        loadRecViewData()
+        getCars()
 
         return autosView
     }
 
-    private fun loadRecViewData() {
+    override fun onStart() {
+        super.onStart()
+    }
+
+    private fun getCars() {
+        val service = APIServicesBuilder.create()
+
+        service.getCarList().enqueue(object: Callback<List<Car>> {
+            override fun onResponse(
+                call: Call<List<Car>>,
+                response: Response<List<Car>>
+            ){
+                println(response.body()!!)
+                loadRecViewData(response.body() ?: emptyList())
+            }
+            override fun onFailure(call: Call<List<Car>>, t: Throwable) {
+                println(t.message)
+            }
+        })
+    }
+
+    private fun loadRecViewData(list: List<Car>) {
         autosRecView.apply {
             layoutManager = LinearLayoutManager(requireActivity())
-            adapter = CarsAdapter(autosViewModel.getCarList())
+            adapter = CarsAdapter(list)
         }
     }
