+++ app/src/main/java/com/example/parcialtp3/fragments/CarViewItemFragment.kt
@@ -8,10 +8,6 @@ import android.view.ViewGroup
 import com.example.parcialtp3.R
 
 class CarViewItemFragment : Fragment() {
-    // TODO: Rename and change types of parameters
-    private var param1: String? = null
-    private var param2: String? = null
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
     }
